# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: NexoShop API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /wheel/spin:
    post:
      operationId: spinWheel
      tags: [wheel]
      summary: Spin the wheel of destiny
      responses:
        "200":
          description: Spin result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WheelSpinResult"
        "400":
          description: No spins available
...
components:
  schemas:
...
    WheelSpinResult:
      type: object
      properties:
        reward:
          type: string
        rewardType:
          type: string
        rewardValue:
          type: ["number", "null"]
        message:
          type: string
        newBalance:
          type: ["number", "null"]
        newPoints:
          type: ["integer", "null"]
      required: [reward, rewardType, message]
...